requirements.txt
typer[all]
python-whois
dnspython